Extract model selector from NewPredict
