+++ src/pages/NewPredict.jsx
@@ -1,29 +1,11 @@
 import React, { useState } from "react";
-import { useQuery } from "react-query";
-import { getModelList } from "../api/model";
+import useModelSelector from "../components/useModelSelector";
 
 function NewPredict() {
   const [predictDate, setPredictDate] = useState("");
   const [predictTime, setPredictTime] = useState("");
-  const [modelId, setModelId] = useState(0);
-  const {
-    data: models,
-    isLoading,
-    isError,
-    error,
-  } = useQuery("model-list", getModelList);
-  if (isLoading) {
-    return <p>正在載入模型列表...</p>;
-  }
-  if (isError) {
-    console.log(error);
-    return <p>發生錯誤</p>;
-  }
-  const options = models.map((model) => (
-    <option key={model.id} value={model.id}>
-      {model.name}
-    </option>
-  ));
+  const { modelId, ModelSelector } = useModelSelector();
+  console.log(modelId);
   return (
     <div>
       <h1 className="big-title">填寫預測數據</h1>
@@ -47,9 +29,7 @@ function NewPredict() {
           required
         ></input>
         <label style={{ marginLeft: "10px" }}>選擇模型: </label>
-        <select value={modelId} onChange={(e) => setModelId(e.target.value)}>
-          {options}
-        </select>
+        <ModelSelector />
       </form>
     </div>
   );
